Use 4-byte header for string length.
git-svn-id: svn+ssh://svn.java.net/grizzly~svn/branches/2dot0/code@6090 04fdb32c-e91d-0410-84f5-98e489d73e21

+++ modules/grizzly/src/main/java/org/glassfish/grizzly/utils/StringDecoder.java
@@ -137,7 +137,7 @@ public class StringDecoder extends AbstractTransformer<Buffer, String> {
                 return TransformationResult.createIncompletedResult(input);
             }
 
-            stringSize = (int) input.getShort();
+            stringSize = (int) input.getInt();
             lengthAttribute.set(storage, stringSize);
         }
